Changes to folder structure
