+++ api/src/main/java/com/uade/api/dao/EspacioComunModelImpl.java
@@ -1,6 +1,6 @@
-package com.uade.api.models.DAO;
+package com.uade.api.dao;
 
-import com.uade.api.models.Entity.EspacioComunModel;
+import com.uade.api.models.EspacioComunModel;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import org.hibernate.Session;
