+++ api/src/main/java/com/uade/api/dao/IEspacioComunModelDAO.java
@@ -1,6 +1,6 @@
-package com.uade.api.models.DAO;
+package com.uade.api.dao;
 
-import com.uade.api.models.Entity.EspacioComunModel;
+import com.uade.api.models.EspacioComunModel;
 
 import java.util.List;
