+++ api/src/main/java/com/uade/api/models/TipoUsuario.java
@@ -1,4 +1,4 @@
-package com.uade.api.models.Entity;
+package com.uade.api.models;
 
 public enum TipoUsuario {
     admin, duenio, inquilino
