+++ api/src/main/java/com/uade/api/models/UnidadModel.java
@@ -1,20 +1,25 @@
-package com.uade.api.models.Entity;
+package com.uade.api.models;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 @Entity
-@Table(name = "espacios_comunes")
+@Table(name = "unidades")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class EspacioComunModel {
+public class UnidadModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idEspacioComun;
+    private Long idUnidad;
     private int piso;
-    private String descripcion;
+    private int numero;
+    private boolean habitada;
+    @OneToOne
+    private UsuarioModel duenio;
+    @OneToOne
+    private UsuarioModel inquilino;
     @ManyToOne
     @JoinColumn(name = "Edificio_Id")
     private EdificioModel edificio;
